Enables shortcode parsing for terms
