+++ admin/taxonomy/class-taxonomy.php
@@ -164,7 +164,8 @@ class WPSEO_Taxonomy {
 							'scope'                    => $this->determine_scope(),
 						],
 						'shortcodes' => [
-							'wpseo_shortcode_tags' => $this->get_valid_shortcode_tags(),
+							'wpseo_shortcode_tags'          => $this->get_valid_shortcode_tags(),
+							'wpseo_filter_shortcodes_nonce' => \wp_create_nonce( 'wpseo-filter-shortcodes' ),
 						],
 					],
 					'worker'  => [
